+++ backend/Application/Api/GraphQL/Metrics/BlockMetrics.cs
@@ -13,8 +13,10 @@ public record BlockMetrics(
     double? AvgFinalizationTime,
     [property:GraphQLDescription("The current total amount of CCD in existence.")]
     long LastTotalMicroCcd,
-    [property:GraphQLDescription("The current total CCD released according to the Concordium promise published on deck.concordium.com. Will be null for blocks with slot time before the published release schedule.")]
+    [property:GraphQLDescription("The total CCD Released. This is total CCD supply not counting the balances of non circulating accounts")]
     long? LastTotalMicroCcdReleased,
+    [property:GraphQLDescription("The current total CCD released according to the Concordium promise published on deck.concordium.com. Will be null for blocks with slot time before the published release schedule.")]
+    long? LastTotalMicroCcdUnlocked,
     [property:GraphQLDescription("The current total amount of CCD in encrypted balances.")]
     long LastTotalMicroCcdEncrypted,
     [property:GraphQLDescription("The current total amount of CCD staked.")]