+++ backend/Application/Program.cs
@@ -2,7 +2,6 @@ using System.IO;
 using System.Net.Http;
 using Application;
 using Application.Api.GraphQL;
-using Application.Api.GraphQL.Bakers;
 using Application.Api.GraphQL.EfCore;
 using Application.Api.GraphQL.Import;
 using Application.Api.GraphQL.Import.Modules;
@@ -52,6 +51,13 @@ builder.Services.AddSingleton<IFeatureFlags>(featureFlags);
 logger.Information("Feature flag [{name}]: {value}", nameof(featureFlags.MigrateDatabasesAtStartup), featureFlags.MigrateDatabasesAtStartup);
 logger.Information("Feature flag [{name}]: {value}", nameof(featureFlags.ConcordiumNodeImportEnabled), featureFlags.ConcordiumNodeImportEnabled);
 logger.Information("Feature flag [{name}]: {value}", nameof(featureFlags.ConcordiumNodeImportValidationEnabled), featureFlags.ConcordiumNodeImportValidationEnabled);
+
+var nonCirculatingAccounts = builder
+    .Configuration
+    .GetSection("NonCirculatingAccounts")
+    .Get<List<string>>()
+    .Select(str => new ConcordiumSdk.Types.AccountAddress(str).GetBaseAddress());
+builder.Services.AddSingleton<NonCirculatingAccounts>(new NonCirculatingAccounts(nonCirculatingAccounts));
 
 builder.Services.AddMemoryCache();
 builder.Services.AddCors();